Move CollectionsModal to src/shared
