+++ packages/app-project/src/shared/components/CollectionsModal/CollectionsModal.spec.js
