+++ packages/app-project/src/shared/components/CollectionsModal/CollectionsModalContainer.spec.js
