+++ packages/app-project/src/shared/components/CollectionsModal/components/CreateCollection/CreateCollection.spec.js
