+++ packages/app-project/src/shared/components/CollectionsModal/components/CreateCollection/locales/en.json
