+++ packages/app-project/src/shared/components/CollectionsModal/components/SelectCollection/SelectCollection.js
