+++ packages/app-project/src/shared/components/CollectionsModal/components/SelectCollection/index.js
