+++ packages/app-project/src/shared/components/CollectionsModal/components/SelectCollection/locales/en.json
