+++ packages/app-project/src/shared/components/CollectionsModal/index.js
