+++ packages/app-project/src/shared/components/CollectionsModal/locales/en.json
